use cp-kafka instead of cp-server
reasoning: cp-server is licensed under the Confluent Commercial License
and it requires a subscription to use it. cp-kafka is licensed under
the Confluent Community License. Technically it's not an OSI-approved
license either, but most users can use it freely.

I also removed the metrics reporter configuration as this is not part
of the cp-kafka image.